+++ src/main/java/org/socialculture/platform/performance/dto/response/PerformanceDetailResponse.java
@@ -1,8 +1,8 @@
 package org.socialculture.platform.performance.dto.response;
 
 import lombok.Builder;
-import org.socialculture.platform.performance.dto.CategoryContent;
-import org.socialculture.platform.performance.dto.PerformanceDetail;
+import org.socialculture.platform.performance.dto.domain.CategoryContent;
+import org.socialculture.platform.performance.dto.domain.PerformanceDetail;
 import org.socialculture.platform.performance.entity.PerformanceStatus;
 
 import java.time.LocalDateTime;
