+++ src/main/java/org/socialculture/platform/performance/repository/querydsl/PerformanceRepositoryCustomImpl.java
@@ -5,9 +5,9 @@ import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import org.socialculture.platform.member.entity.QMemberEntity;
-import org.socialculture.platform.performance.dto.CategoryContent;
-import org.socialculture.platform.performance.dto.PerformanceDetail;
-import org.socialculture.platform.performance.dto.PerformanceWithCategory;
+import org.socialculture.platform.performance.dto.domain.CategoryContent;
+import org.socialculture.platform.performance.dto.domain.PerformanceDetail;
+import org.socialculture.platform.performance.dto.domain.PerformanceWithCategory;
 import org.socialculture.platform.performance.entity.PerformanceStatus;
 import org.socialculture.platform.performance.entity.QCategoryEntity;
 import org.socialculture.platform.performance.entity.QPerformanceCategoryEntity;
